[fidl][dart] Guard against null/missing fields
This changes the generated struct init constructors to take field names
as named params and marks the non-optional ones as @required. It also
checks that the non-optional fields are not null.

I have trivial changes to Topaz in my tree that fixes where init
constructors are called directly.

Next we could see if the default zero args struct constructor can be
replaced by the init constructor.

Change-Id: I940ce0bca0f2175a1c2fc92e203ca211a6f75754