[fidl][wlan] Introduce WlanChan struct
This is to replace the previous use of uint16/uint8 channel
in reporting the BSS channel / setting the radio's channel.
This is to be used to accommodate the channel bandwidth support.

Note, WlanChan is not for channel capability report; there
the radio reports the list of frequencies it can tune to, in channel number.
It shall remain as int8 list.

NET-319

Change-Id: If2081b5147062d863a8e6d2624f47f76a03c8c48