[gpt] unify gpt structure headers
The bootloader wants to see these, so migrate to system/zircon/hw
and try to reduce the number of slightly different versions of these
structures and defines we h ave.

ulib/gpt will need more invasive changes to use the common version
of the gpt header that does not make assumptions about the block size,
so I renamed its gpt_header to gpt_hdr_blk to more accurately reflect
what it is, to avoid conflicts until more extensive tidying up can
be done.

Change-Id: I4471ecd6332cc8d071e0bced4df900ccf2579e99