[fs-vnode] Shutdown message loop before calling VFS destructor
This test was triggering an ASAN use-after-free violation.

The fs::Vfs object had handlers registered to an async
message loop, but was deleted before the message loop
removed all handlers. As a consequence, when terminating
the message loop connections, the async::Loop destructor
dereferenced the vfs object which had already been deleted.

Change-Id: Idb3a5572932346a3775a2e8c1227ac821ec98be3