[netcp] Treat ENOBUFS as EAGAIN and EWOULDBLOCK for macOS’ benefit
On macOS and other BSDs, poll() doesn’t reflect whether the interface
has buffers available to service a send(). The commonly used tuntaposx
driver (http://tuntaposx.sourceforge.net/) uses a QUEUE_SIZE of 128,
which doesn’t interact well with system/host/netprotocol/netprotocol.h’s
TFTP_DEFAULT_WINDOW_SZ of 256. Attempting to send() when no more mbufs
are available in the interface driver results in ENOBUFS, which shows up
in netcp as frequent “netcp: failed to handle message (status = -1)”
errors when attempting to send a large file to an emulated QEMU
instance.

A smaller TFTP_DEFAULT_WINDOW_SZ might be appropriate on macOS, but
since the state of the interface’s queue is unknown to netcp, and since
this similar to the EAGAIN and EWOULDBLOCK situation addressed by a
retry in 163e67a0cd13b, ENOBUFS is added to the set of errors that a
send() can be retried for.

Change-Id: If4609f656e0283dc3b548f6f6ae2453b7852578f