[tests] Fix race condition in link_context_entities test.
The race condition was that the LinkWatcher in the story_info agent was
registering at an unknown time. In this time, the Module writing to its
Link may or may not have already started mutating Link state.

The test was relying on the fact that the Link state changed from one
value to another, which is an abuse of Links: they are there only to
hand clients the latest state, and make no guarantee about exactly when
the clients begin being notified of state changes.

This fix simply tests the two conditions we wanted to test using two
different Links. We now check that both values from both Links propagate
into and out of the context engine correctly. It works because we're
guaranteed to eventually see the right state - we just have to wait long
enough.

MI4-759 #comment [tests] Fix race condition in link_context_entities test.
MI4-759 #done

Change-Id: Ibfee15875c32403d4bd7637a3f0231e87d6bb2ec