[kernel][syscalls] Automate handling of handle result parameters
System call implementation functions now get a `user_out_handle*`
instead of a `user_out_ptr<zx_handle_t>` for result parameters
that are handles.  The sysgen-generated code and the new class
replace the boilerplate code for instantiating handles and copying
out the zx_handle_t values to the user.

Change-Id: I34bbc68d198d6933506c20a76b27e2586cd1ce8d