[virtio] RingKick should be based on ring_index
RingKick was incorrectly using the information of the last configured
ring to kick. The kick is based on the address written to, not the value
written. The address written to is based on the ring_index, not the last
configured ring.

Without this change, the last ring to be configured would be the ring
that all kicks would be directed to.

Change-Id: If9a5e05d281cc765d2bf353cb7faefac09ecc5d6