[fidl] Introduce PreallocMessage for stack-allocating small messages
Use PreallocMessage for receiving messages via ReadMessage().  This
avoids the need for a malloc()+free() for small messages.  This
doesn't yet reduce the number of zx_channel_read() calls -- I'll do
that in the next change.

Note that the choice of size for prealloc_buf_ (128 bytes) is fairly
arbitrary at the moment -- I haven't tried tuning it.

Also note that I haven't used PreallocMessage for constructing
messages to write.  For that case it will be better to allocate an
exact-sized buffer.

US-401

Change-Id: Iddcd65b04a893e5c95fb5685111cfe7fcfc8c0b5