[fidl] Change AllocMessage's MoveTo() method to MoveFrom()
This is for consistency with the recently-added CopyDataFrom() and
MoveHandlesFrom() methods.  It's also more consistent with the
direction in which std::move() is used, though I didn't go as far as
converting this to use an "&&" reference.  This is currently only used
by test code, but it seemed worthwhile making it consistent.

US-401

Change-Id: I8194ab697cc26591354f868aba60ad3736f3b2b4