[maxwell] Using agent controller to detect crash
Agent runner closes the agent controller connection when the agent
terminates. We restart Kronk (up to a limit) when we notice this.

Although the interface we're actually interested in is kronk_services_,
we still need to put the restart handler on the controller. When the
agent crashes, kronk_services_ often gets closed quite a bit earlier (~1
second) than the agent runner notices via the application controller
(which it must use as opposed to any interface on the agent itself since
the agent is not required to implement any interfaces itself, even
though it is recommended that it does). If we try to restart the agent
at that time, the agent runner would attempt to simply send the
connection request to the crashed agent instance and not relaunch the
agent.

MI4-771 #done

Change-Id: Ie170a7af673da9cd86065db2285896733c224f33