[vim2] Add Khadas VIM2 build
The VIM2 is very similar to the original VIM board,
but has an Amlogic S912 instead of the S905X.

Kernel configuration is identical, except VIM2 has a second CPU cluster
and a different set of reserved ranges.

system/dev/board/vim2 is identical to system/dev/board/vim,
except the framebuffer address for the display driver hack is different.
We will refactor the board drivers to avoid code duplication in a subsequent change.

Change-Id: I65bd2620cf261975fa0854ef48f4e3f214e00e7e