[modular][ledger] unify the implementations of TestWithMessageLoop.
This patch merges the forked implementations of the TestWithMessageLoop
test base class into one, to be used by unittests across peridot. (this
has originally been the same code, copied over between Ledger and
Modular before the repos got merged)

The new class lives in a separate lib/gtest library. (In follow-up
patches we can continue to move things out of lib/testing which is a bit
of a dumping ground now)

Change-Id: I4d80c2269ad15daeb28b48f76f10fad6676244fa