[wlan] FrameHandler clean-up.
Introduces a new internal methods for each frame type.
This cleans up the code but is also required to further extract the
frame forwarding logic out of the FrameHandler and into a dedicated
class such that forwarding of frames is more performant and can be
directed towards a specific child rather than broadcasting the frame to
everyone.

Change-Id: I022b9b050f22bd2c424587b9c564561d8d1de44d