[system][hid] model generation
Now the hid parser finally generates the output
model and with that a lot of detailed testing
is done.

The testing is a "bit" tedious but bit level
testing is done for:
- boot mouse
- adafruit composite device (4 in 1)
- ps3 joystick

Also a usage header is introduced with hundred
of usages, with keyboard missing because there
is already one for that purpose that needs
fixing.

Change-Id: I7f8cadb3d36f61eb62827e6bd3de8fb8df1e7b28