Testing best practices
The purpose of this doc is to explain concepts that I repeatedly explain
to Fuchsia developers, about defining our goals when writing tests and
figuring out how to write tests that achieve those goals.

The structure I'm going for is to first define an "ideal test" by
example. The ideal test is what we want tests to look like, and I
explain why. Then, I show how we can take a messy real-world Fuchsia
test and nudge it toward the ideal, based on an actual CL I wrote.

There's quite a bit of stuff that I left out of this doc because I don't
want it to be overloaded with "tips" or a list of "bad practices" versus
"good practices". My idea is that it's more effective to get a single
concept across and then let people work out for themselves how to work
with that concept. This also matches the way I've been working with
individual developers on test techniques: I rarely just say something
like, "This technique is bad". Rather, I point out ways in which a
certain technique succeeds/fails at our goals.

Questions for reviewers:
- Does the concept of "ideal test" make sense to you as a basis for
  reasoning about test strategy?
- Is the example ideal test a good one? Is it too simplified, or
  unnecessarily complex? Is it missing some important aspect that you
  think should be illustrated?
- Is the example real CL well-connected to the concepts?

Change-Id: I4d00bcbf03541f5fd0fa3bb0db40490694ff2a94