[appmgr] Retain file system that backs /pkg
When an ApplicationRunner is responsible for controlling an application,
we previously shut down the file system that was backing /pkg for the
application. Now we keep it alive with the runner.

This CL keeps the file system alive a bit too long (for the lifetime of
the runner rather than the lifetime of the application being run). The
lifetime issue will be resolved when the package manager is responsible
for hosting these file systems.

Change-Id: I83437ecc7380b580050cc8e1f7683c1e8cb6db68