cobalt: Stop setting verbose=10 on the CI bot.
We had temporarily enabled gRPC debug logging on the CI bot
but the problem we were trying to diagnose has mysteriously solved
itslef so we don't need this anymore. Note that it's not the case
that the verbose logging is the cause of the fix because the
CI turned green at 2:00 p.m. and the CL to enable debug logging
was merged at 4:00 p.m.

Change-Id: I28290772c3323bd7c41fb5c809734d52d2579ebe