[runtests] Emit summary JSON file with -o.
This change makes -o emit a JSON file containing a summary of
test results, and a JSON schema describing the structure of the
summary. This provides a machine-readable way to parse test results,
rather than relying on just the output files, since several different
test frameworks are used throughout the Fuchsia stack.

Change-Id: I0994fd8e495250d7acf3f451e16daeaecaf5750d