[disk-pave] Skip inappropriate x86 partitions
On a CROS device, avoid writing an EFI partition.
On a non-CROS device, avoid writing a KERNC partition.
On an unknown device (e.g. blank disk), avoid writing a KERNC partition.
If --force is given, write whatever is requested.

Rule 3 for unknown devices has a subtle importance:
When bootstrapping a blank EFI system (e.g. a NUC) for the first time,
we could instead write everything, however, doing so would then result
in the device being identified as a CROS system for future updates -
effectively it would then never update the EFI partition. Instead we
currently make the assumption that is_cros() provides accurate results
for all CROS devices, which it does for systems we've tested.

Change-Id: Ie7226bcc99562dfef3db898f07602832ecfda074