[ddk][block] new block protocol
The plan here is that we will use a protocol more tightly aligned
with the needs of the block IO layer (including constraints that
make things much more consistent and less open ended than the
more free-form iotxns did).

The expectation is that the only client of this protocol will be
the generic block mid-layer driver.

The deployment plan is:
1. Agree on the new protocol
2. block mid-layer driver will support new protocol, if present
3. block drivers will support new protocol
4. once all block drivers support new protocol, we retire iotxn_queue()

Change-Id: I2659fe985f1a815352db67faa3788fb3c239ea66