[arm64][hypervisor] Save guest virtual timer
Make sure we save the guest the virtual timer on VM exit. We then rely
on these saved values for timer handling, rather than on the current
value of the timer registers which may have been modified by the host.

Furthermore, this CL restructures the El2State and modifies the handling
of HostState in el2.S, so that we can increase the size of El2State
without hitting the limits of immediates in instructions.

Change-Id: I110397094be4f40bdce1ced25390554415c5d1c3