[kernel][hypervisor] Do not require INVPCID support
INVPCID is an optimization that allows to avoid flushing TLB, it
should be possible to run hypervisor without VMX support for INVPCID.
Particularly when using KVM in nVMX mode, nested VMX doesn't support
it, at least in not bleeding edge KVM versions.

This change makes VMX INVPCID support optional and clears INVPCID flag
in CPUID if VMX doesn't support it, thus proper guests shouldn't use it
in this case.

Change-Id: Ide1c2da57471f6c223c9888ecf7eacbd87a7aae8