[uapp][threads] Don't "hang" if thread is terminated.
The program won't really "hang", it will timeout, but the timeout is
a bit large so that it works on really loaded systems. Normally waiting
for the thread to suspend will happen quickly, and the thread could die
not infrequently so don't give the appearance of hanging in that
not-infrequent case.

Also, drop the timeout from 10 seconds to 5 seconds.
Dunno what's a good value here, 10 felt a little long.

Change-Id: Ibd73b312f739c3aca2d8ffd8b42e0f451bee6d59