[virtcon] Add support for hotplugging
This change adds support for hotplugging to the virtcon. If multiple
displays are connected, the virtcon will use the display which was
attached first. The virtcon's contents and scrollback are kept when
switching between displays, although long lines will be truncated if
the virtcon moves to a smaller display.

When no display is connected, the console continues to function
normally, except without rendering anything. All checks for whether or
not any gfx operations should be done are in vc-device.

Change-Id: I467b9b1792f4d43803b521ec564dfe1873b85a97