[packages] Hoist labels up the gn deps chain
Previously, labels were deps of the step that created the boot
manifests.  However, now that we use the package() template to pull
build products into boot manifests, we don't need the targets specified
by the labels to build built at that stage.

This CL moves the labels up the dependency to tree to be deps of the
default target, which will let us include system images as labels.

Also, group together the targets needed for bootfs to make their gn
label easier to spell.

Change-Id: Iebb6a1765dd2197cbd2d5664378976479319ce95