[tftp] Add support for BUSY response to requests
If a server is unable to handle a request at the moment, but may
be able to do so if the client waits, it can respond with a
TFTP_ERR_CODE_BUSY error packet. This behavior is induced by a
response of TFTP_ERR_SHOULD_WAIT from a file_open_read() or
file_open_write() call.

ZX-1521

Change-Id: I2551407052d4a33a7e6b6eebcf45d8f4b28b9dab