[runtests] Don't build output path for every test.
This change hoists out the code for generating the part of the
output path that is independent of the test binary's name, since
it's unnecessary to do this every time. It also resolves the
path to the test binary using realpath to completely eliminate
any chance of output name collisions via relative paths (e.g.
runtests -o /data /system/test ./system/test).

Change-Id: Ibe014745944df3e836b772564dff6c36c2469728