[pm] cleanup the meaning of using a manifest
If a manifest is supplied, it must specify all of the files to be included in
a package with the only exception of the generated meta files
(meta/{contents,pubkey,signature}). (In other words with must now include
meta/package.json). If a manifest is supplied the output dir is NOT used to
gather package contents.

Change-Id: If0c9fb9d11bb6795191f7f8830bd8d1e0fbad62f