[framebuffer] Fix double locking issue
The magma implementation of acquire_or_release_display makes a call to
the ownership callback, which takes a lock. Change 866e4bc26 made sure
that the framebuffer's parent device is never used after unbinding, and
in doing so it moved a call to mtx_lock before calls to FB_ACQUIRE and
FB_RELEASE, resulting in a double locking issue. Since the magma driver
never unbinds, it isn't necessary to move the lock for that driver, and
the framebuffer driver will go away before more GPU drivers are added.

Change-Id: I1790ec476420640a7453fe6bcca0f22a9932eb40