[machina] Remove references to LocalApic
The IO APIC now registers VCPUs instead of LocalApics. Now that we've
moved to x2APIC, logical interrupt addressing is based off the local
APIC ID so we don't need LDR or DFR. In theory, x2APIC logical
addressing should take into account the cluster ID as well but the
IO APIC has no notion of clusters. We'll need better handling for this in the
future.

Change-Id: Ia1b28867e8fb8c1cdf8e1c39f84dad43e67167f3