[modular] Expose the TokenProviderFactory to Ledger
This is a stopgap solution while we have a proper transport layer
that can do authentication. Pending this layer and for the
peer-to-peer ledger work, we need a way to uniquely identify users
to avoid users clobbering each other. TokenProvider's client_id will
be used for that.

This CL plumbs TokenProviderFactory to the Ledger (and only the Ledger).
This is meant to be removed once we have a better solution.

Change-Id: I42e28bc0c292b6b2347f2b712ad640895d6ecfb5