[fidl] Add fidl::String
This patch starts to introduce the C++ FIDL types. These types are views
into the underlying C types that are used by the encode and decode
functions. For decoding, these views are ultimately on the underlying
message buffer, similar to the C types. For encoding, the underlying C
types are stored in a builder and manipulated through the C++ views.

Change-Id: I21b8e2be687c2cfed57d395898e31e3d16bc9961