[unittest] Add ASSERT_NO_DEATH
Add the ability to run a function and assert that it did crash in
addition to the ability to assert that it did.  This lets me write
tests where I have an operation which might crash (because the
code-under-test is broken), but should not, and still terminate the
test cleanly with a clean ASSERT log message after catching the
exception.  Afterwards, other tests in the test case can continue to
run even if this specific test failed.

Change-Id: I7778c07cba16233252d20de1dd9c9bca9fa960a0