[kernel][hypervisor] Handle KVM features CPUID leaf
Zircon hypervisor disguises itself as KVM, which makes Linux kernel
believe that kvmclock is available and try to use it (see
kvmclock_init function in /arch/x86/kernel/kvmclock.c in the Linux
kernel code).

This CL changes CPUID handling so that it reports all KVM features
as unsupported, at least for now.

Change-Id: I081a82df9242e47d63ebd11bbb400f17c588a7e4