[block] Fix intermediate error support for block transactions
Previously, certain errors would call an "out-of-band" error
response function if a response was requested. Unfortunately,
in the case of a multi-message operation, if any intermediate
messages caused an error, they would be silently dropped.

This patch addresses that issue by ensuring that for ALL
requests which have a valid txid, errors are "remembered"
and transmitted when a response is requested.

ZX-1581 #done

Change-Id: Ife2f91f6f756bb8333664c6f234cf12f31cbab16