[ledger] Save identifier information for unsynced objects.
This changes the on-disk database such that:
- Object data continue to be keyed by digest
- Object sync state is now keyed by object identifier.
In particular, it means that the same object data can has different sync
state depending on its key index and deletion scope. To allow this, it
was not possible to consider that any object in the disk database that
had no sync state was synced, so this CL change the database to
explicitely contain the synced state.

This CL also removed DeleteObject from page_db.h, as it was never used
and is not trivial to implement when an object can be referenced
multiple time with different identifiers.

LE-286

Change-Id: I1e506e1b8a3ed6532cc37fa2717fab32f2db0e0b