[fidl] Add fidl::Message data structure
The fidl::Message data structure holds both the byte and the handle part
of a FIDL message and knows how to encode and decode messages in place.
The fidl::Message also knows about the message header and its various
fields.

This CL also moves the C++-related headers into a cpp directory to make
it easier to see which parts of the library are related to the C++
bindings and which are useful from C.

Change-Id: I7031d1b3d9a76b911f7e9de9c3f5aa33a0f700f8