[kernel][thread] Reorder operations in thread_suspend()
Set the signal in t->signals before we unblock the thread.

This is in preparation for moving the sched_reschedule() call into
wait_queue_unblock_thread() and/or sched_unblock().  We want to modify
t->signals before the target thread gets to run, otherwise the thread
suspension might fail to take effect.

ZX-1490

Change-Id: If14d2b791cf1d891f151acf06767b0e0db95a7df