[crypto] Refactor tweaked codebook Ciphers
This CL makes tweaked codebook Ciphers more explicit and simpler to use.
Generic Init and Transform methods are exposed with an explicit
direction and the InitEncrypt/InitDecrypt and Encrypt/Decrypt aliases
have both random access and stream cipher versions.  For random access
mode, the "tweakable" mask has been replaced with an "alignment" field;
allowing the tweak to be calculated automatically from the offset.

Change-Id: Ifd1eef6dd194f0131da099da14962437486770b3