[resolver] Teach ModuleResolver to set create_chain_info.
This is one step in the process of instructing the Framework to set up a
Module's Links correctly for Modules using their manifests.

There are a number of next steps:
* Create a new structure to serve as the input type for the
ModuleResolver, replacing Daisy. This allows Daisy to stay relevant only
to the abstraction layer between Modules <-> Framework (where LinkPaths
don't exist as a concept).
* Use this new structure as ModuleResolver's input, and include on it a
Link snapshot instead of a LinkPath. This allows the ModuleResolver to
avoid knowing how to get from a LinkPath -> Link data.

This will allow ModuleResolver tests to avoid taking dependencies on
StoryProviderImpl and StoryControllerImpl, as both of those are
currently required to operate code that wants to resolve LinkPaths to
Link instances.

MI4-725 #comment [resolver] Teach ModuleResolver to set create_chain_info.
MI4-725 #done

Change-Id: Ib42c54da98e950391a2d96beccfe4f08f77b4955