[resolver] Split Daisy into two structs.
One is for the interface between Modules and the Framework, and another
is for the interface between the Framework and the ModuleResolver.

This sets the stage also for the ModuleResolver to be hosted in the
cloud, as it starts removing ModuleResolver's dependencies on Framework
components by making ResolverQuery closer to a POD type. All that's left
is making the job of extracting type information from an Entity
reference the job of the Framework, instead of the ModuleResolver.

MI4-732 #comment [resolver] Split Daisy into two structs.

Change-Id: Icb23a92539fdeccd0d7b05bac32ddfc2465341a2