[docs][syscalls][interrupt] revised interrupt api
A rearrangement of the interrupt API to make interrupt objects 1:1
with the underlying physical (or virtual) interrupt, and allow any
number of such objects to be bound to a port.

zx_interrupt_create() may be used by privileged DDK code (bus
managers) to create physical interrupt objects, and by any
code to create virtual interrupt objects.

For the most common case of a single interrupt and a single irq
thread dedicated to that interrupt, zx_interrupt_wait() blocks
on the irq (re-arming it) and zx_interrupt_destroy() can be used
to safely signal the irq thread that it must shut down.

For more complex situations, zx_interrupt_bind() allows an interrupt
to be bound to a port, where interrupt packets will be delivered
if it is triggered and zx_interrupt_ack() is used to re-arm the
interrupt.  zx_interrupt_destroy() can be used to case an interrupt
canceled packet to be delivered to request clean shutdown or
disconnect of that interrupt.

zx_interrupt_trigger() is provided to trigger virtual interrupts.

Change-Id: Id0a56e44193b1062e4b3a81989ac3351bec085e4