[fvm] Improve noncontiguous access test
This test was sufficient for preventing the regression
of LE-388 using the old block protocol, where an iotxn_clone
would fail with invalid arguments, but with the new block protocol
the test could pass (incorrectly) if the vmo_off was miscalculated.

This path checks that the complex path of noncontiguous slice
access is validated against contiguous slice access reads.
This improvement prevents the regression, even with the incoming
new block protocol.

ZX-1578 #done

Change-Id: I034bfe2d90307f19b6ec69659d6ce01fa5feb38e