[arm64][hypervisor] Change how we handle timers
On every VM exit, we check whether the guest timer was set. If so, we
setup a timer object to fire and inject an interrupt when the guest
expected one.

Previously, we would wait for a WFI before we setup a timer object,
which would cause us to incorrect inject a timer.

We also now ensure that interrupts are unmasked before we inject them.

Change-Id: I1a1a2319a6d358f5a86d45ee251bafb6156a22bd