[ledger] Remove calls to zx_cprng_add_entropy
This CL removes the InitEntropy and EnsureInitEntropy functions.
As written, they were not adding predictable data as entropy to
the system CPRNG.  Furthermore, re-seeding the CPRNG is not to be
performed by userspace applications directly.  Ledger and other
applications can zx_cprng_draw directly and assume the platform
has initialized and is reseeding the CPRNG as needed.

Change-Id: Id91829b9c3c671c80affa638f69f5795e7600938