[pci] Remove DEBUG_ASSERT in Irq bookkeeping
ResetCommonIrqBookkeeping is called from both LeaveLegacyIrqMode
and LeaveMsiIrqMode. The existing assert is written to assume
that if you have > 1 irq handlers then you have the handler table,
and otherwise you're using the singleton. However, the singleton path
also catches the case where this is called when you have no configured
handlers. This path is hit if a driver tries to enter MSI mode, then
bails out due to no backing MSI blocks being available in the platform
driver.

It appears safe to remove and refactor since all the work it does is
safe unconditionally.

ZX-1511 #in-progress
ZX-1512 #in-progress

Change-Id: I29eb70a5367df265f2b4b6ed7ecd766f06dc1441