[packages] default to building Fuchsia packages
The configuration variables from //build/gn are factored out, and the
user.bootfs portions moved to their own label set in order to make it
possible to assert that the right kind of packages are being built for the
bootfs. This move will also enable the various images targets to be factored
out into their own packages later.

Change-Id: Ifb2c6060c2086a74c225e69f49ae87b33f1a218b