[kernel][hypervisor][arm64] Remove use of INT_RESCHEDULE in timer callback
A use of INT_RESCHEDULE reappeared after I removed it, but
INT_RESCHEDULE should not be necessary here.

Interrupt() calls two functions: Track() does not unblock any threads,
it only sets a bit in a bitmap.  Signal() can unblock a thread, but it
uses reschedule=true, so we don't need to return INT_RESCHEDULE as
well.

ZX-1490

Change-Id: Ibf41f7074bf26bb4df15238a40225d0603b49a5c