[kernel] Convert remaining timer callbacks to make use of preempt_disable
Remove the remaining uses of INT_RESCHEDULE in timer callbacks.
Convert to calling thread_preempt_set_pending() or passing
reschedule=true.

Note that for threadload() in debug.c, printf() currently does not
trigger a reschedule by itself for running debuglog.  In this CL, I am
not changing printf() to trigger a reschedule yet, because that has
potential to cause performance regressions.

ZX-1490

Change-Id: I2b5f13028fdcb1ca629ae57ac9dda86ede2505d2