[virtio][block] convert to new block protocol
This works with the iochk exerciser, but does not pass the minfs
test.  However the pre-conversion version fails the minfs test the
same way: if requests arrive fast enough there may not be enough
ring entries for a request and the driver does not have a mechanism
to wait until they become available.

Change-Id: I0d92774efbdf1ae1caa1d7359b6a3084047e7494