[ramdisk] Remove a race-prone MAX_TXN_MESSAGES test
The "too many ops" test attempted to:
 - Write MAX_TXN_MESSAGES to the block server (automatically appends
   "we should respond to the client" identifier on the server).
 - Write another message to the block server
 - Observe that ZX_ERR_IO occurred, because too many messages
   were enqueued (message seen after "we should respond" marker).

Unfortunately, this test is racy. It's possible that the first
group of messages completes and responds to the client (returning
a status of ZX_OK) before the next message (attempting to prod
ZX_ERR_IO) is even enqueued.

Removing this test, since the MAX_TXN_MESSAGES limiter is being
removed in v2 of the block midlayer anyway.

ZX-1600 #done

Change-Id: I0448cbd329b27cb1325d211ea350ade8cf9c5f3c