[netstack] Split ioctl_netc_get_if_info into multiple calls.
This is an expedient fix while transitioning to the new socket control
transport instead of FDIO. Socket control planes only support a payload
of 900 bytes, which does not fit the original ioctl. Rather than a
single ioctl to query the full list of network interfaces, I provide
an ioctl query interfaces one by one.

This ioctl, and the one it replaces, are temporary, until we can
implement getifaddrs - which depends on FIDL support in zircon.

Change-Id: I422b80f096288d857c34163de40805f9b2ff9c53