[mdns] factor out interface enumeration
The interface enumeration code was previously part of the
MdnsTransceiver implementation. This is an issue for building a
standalone (fidl-free) mDNS app, because MdnsTransceiver is
required, but the fidl-based interface enumeration code can't be
part of the app. This CL factors interface enumeration into a
separate class (FidlInterfaceMonitor) that conforms to a fidl-free
abstract base class (InterfaceMonitor).

This CL also fixes a long-standing bug in the mDNS code that would
have prevented it from working with V6 interfaces. Specifically,
MdnsTransceiver used to assign its own indices to network interfaces
rather than using the indices defined by the OS. This is a problem
for V6, because some V6 socket options refer to network interfaces
by index (rather than IP address, as in the V4 case).

Note that InterfaceMonitor and MdnsTransceiver store objects indexed
by interface in vectors, assuming that interface indices will be
densly clustered near zero. This is consistent with the way posix
works. Oddly, the fidl and ioctl interfaces currently exposed by
netstack use 'index' or 'id' fields in their interface descriptors
which disagree with the actual indices of the descriptors in their
respective arrays. This is surprising and will likely be the source
of some frustration as developers accustomed to using posix interface
enumeration use the netstack APIs. mdns_service will have problems if
netstack starts assigning large indices/ids to interfaces. This issue
will be raised with the netstack developers.

Change-Id: I8eb5db8aa1f37bd9dea344832574dc82b3db5291