[ethernet] Read from rx fifo in groups
Rather than read a single rx entry every time we need one, read a
group at a time. This reduces the number of syscalls we make and
will make it easier to recognize when we are running short on
rx entries, which will be useful if we ever want the ability to
apply backpressure onto the device driver.

NET-402

Change-Id: I4ac18f3667b46db6163784bd9477eb63fc1f0efb