[virtio][block] cope with temporary resource exhaustion
Don't fail if all 32 requests are in flight or there are
not enough free descriptor slots available.  Instead wait
and try again as in-flight requests complete.

This could probably be done more cleverly, but for now
I avoided getting tricky.

Also, access alloc/free_blk_req() and the txn_list under
a lock, to avoid the irq thread and queueing threads
messing with each other.

Change-Id: I1d7f6e6e772c665514cddccd867520f56a3002b4