[utest] Check whether the hypervisor is supported
Adds a check to see whether the hypervisor is supported before running
Guest Physical Address Spaces tests on ARM64.

On ARM64, we need to jump to EL2 during operation of the Guest Physical
Address Space. If we run these tests without access to EL2, we'll fail.
This can be seen when run under QEMU without virtualisation support.

ZX-1614 #done

Change-Id: I345af9c09361e64aee4cd58d07385d5228783d2d