[kernel][arm64] make sure we properly account for the ramdisk physical address when booted from zedboot
Any loaders that just pass us a raw pointer to the bootimage in x0 will
go through one of the three paths at the top of platform_early_init to
recover the size of the ramdisk. One of the paths wasn't properly
accounting the ramdisk physical base and length, which was causing newer
boot reservation code to bail and stop the kernel boot.

Only really zedboot and the machina hypervisor pass things this way.

Restructure the code a bit to make sure the physical run of the ramdisk
is always computed.

ZX-1612 #done

Change-Id: Iafd693a901ed7b6dd8dd8e5bb501961eb4eff20d