[utest] Run kernel unit tests ("k ut all") from runtests
Run the in-kernel unit tests on the bots (on the CQ etc.), via
runtests, by adding a userland test program that does the equivalent
of "k ut all".

The kernel-side code propagates enough error results through that this
will detect test failures such as the failure of an EXPECT_EQ(), and
of course it will detect kernel panics too.

IN-19 #done

Change-Id: I949603339df83b04201bde3ac5126c73c4f4822a