[bt][hci] Remove flaky HCI_LowEnergyConnectorTest
HCI_LowEnergyConnectorTest.SendRequestDeleteAndWaitForTimeout is a
poorly written test case that a) depends on racy timing to validate
multiple invariants; b) is almost identical to SendRequestAndDelete,
which does the same thing but without relying on racy timing.

What this test was really testing is the cancelation of a connection
timeout when a LowEnergyConnector is deleted. Since the cancelation
state could not be obtained from the deleted object, the test waited
for the timeout to expire.

The timeout cancelation can be more easily verified in a test that
doesn't involve destruction, so that expectation has now been moved
to another existing test (which is written against
LowEnergyConnector::Cancel() rather than the destructor).

Since cancelation is implied by the destructor and since there is
already a non-racy destructor test (SendRequestAndDelete),
SendRequestDeleteAndWaitForTimeout is no more.

NET-394 #done

Change-Id: I6802a874836ab4070b0dbc2dcaa7f66f496e398d