[ramdisk] Reduce test flake using a directory watcher
When creating a ramdisk, the ramdisk controller binds an
instance device to a topological path. However, the tokenized
path (in "/dev/class/block/XYZ") may require a small gap of
time before being created.

The "test_filesystem" test checks that the ramdisk is visible
in both locations (it was created to prevent a regression of
a fairly old bug), but it did not account for the time gap
before appearing. As a consequence, it's possible that the
tokenized name appears to "not exist".

Update the test to use a directory watcher, removing
this flake condition.

ZX-1620 #done

Change-Id: Idaf8ed1db17d46943c51a959f1040b8436233d9c