[kernel][libc] Fix printf()'s formatting of large unsigned integer values
This line was incorrect:

  int digit = (int)n % 10;

Note that n is of type "unsigned long long".

Truncating n to "int" wasn't right because 2^32 isn't divisible by 10.
This could cause wrong numeric values for some of the digits.

Changing n to be signed wasn't right, because "(int)n % 10" will be
negative if "(int)n" is negative.  This could cause characters other
than 0-9 to be printed for the digits.

Change-Id: Ifbf306b252e7b267f126d348c3bb6eaf7c85d4a2