[devhost][devcoord] adjust opcode differentiator
devfs forwards OPEN messages to devhosts over the same channel
used for devcoordinator messages.  Previously the fact that the
high bit was set on DC messages differentiated them from RIO
messages, but the high bit will be used for FIDL system messages.

This moves the differentiator bit to bit 28, adds a static assert
to verify it is non-conflicting at build time, and fixes the
check for RIO_OPEN in devhost.c to check the entire opcode field.

Change-Id: I60d72cf2f35c6b9d87103be2a41edeeea9fecb8f