[cloud_provider_firestore] fix the FirestoreService shutdown.
This patch makes sure we don't delete the FirestoreService before the
CompletionQueue is fully drained - as the completion callbacks of the CQ
are posted on the main thread, we need to make sure we only delete the
client after they are executed.

LE-368 #comment [cloud_provider_firestore] fix the FirestoreService shutdown.

Change-Id: Id72f2ae02a4462e2a709505704e2d1235f746e3e